+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Add contents list heading Welsh translation (PR #881)
 * Enable passing data attributes to attachment components (PR #874)
 * Add potentialSearchAction to the GovernmentOrganization schema (PR #870)
+* Add potentialSearchAction to manuals (PR #879)
 
 ## 16.22.0